+++ src/clj/reports/routes/services.clj
@@ -34,10 +34,15 @@
       (when (empty? filename)
         (throw (Exception. "did not select a file.")))
       (io/copy tempfile (io/file (str dir "/" filename)))
+      (db/create-upload! {:login login :filename filename})
       (-> (response/found "/r/#/upload")
           (assoc :flash (str "uploaded " filename)))
       (catch Exception e
         (layout/render [request] "error.html" {:message (.getMessage e)})))))
+
+(defn get-logins [request]
+  (let [ret (db/get-logins)]
+    (response/ok ret)))
 
 (defn services-routes []
   ["/api"
@@ -47,4 +52,5 @@
    ["/ping" {:get (fn [_]
                     (response/ok {:status 200
                                   :body "pong"}))}]
-   ["/upload" {:post upload!}]])
+   ["/upload" {:post upload!}]
+   ["/logins" {:get get-logins}]])